pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.5.0",
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
]
